# Feature/tg 3967 immutable reducers