--- a/.babelrc
+++ b/.babelrc
@@ -1,3 +1,4 @@
 {
-    "presets": ["es2015"]
+    "presets": ["latest"],
+    "plugins": ["transform-object-rest-spread"]
 }