--- a/.babelrc
+++ b/.babelrc
@@ -1,4 +1,3 @@
 {
-    "presets": ["latest"],
-    "plugins": ["transform-object-rest-spread"]
+    "presets": ["latest"]
 }