--- a/modules/atlas/services/redux/reducer.factory.js
+++ b/modules/atlas/services/redux/reducer.factory.js
@@ -1,6 +1,8 @@
 (function () {
     'use strict';
 
+    let totalReducerTime = 0;
+
     angular
         .module('atlas')
         .factory('reducer', reducerFactory);
@@ -45,10 +47,17 @@
                 printReducers
             );
 
+            const t1 = performance.now();
+            console.time();
             if (angular.isObject(action) &&
                 angular.isObject(action.type) &&
                 angular.isFunction(actions[action.type.id])) {
-                return actions[action.type.id](oldState, action.payload);
+                const result = actions[action.type.id](oldState, action.payload);
+                const t2 = performance.now();
+                const diff = t2 - t1;
+                totalReducerTime += diff;
+                console.log(`${action.type.id} ${diff} ms,\ttotal: ${totalReducerTime} ms`);
+                return result;
             } else {
                 // TODO: Redux: throw error
                 return oldState;