--- a/modules/atlas/services/routing/state-url-converter.factory.test.js
+++ b/modules/atlas/services/routing/state-url-converter.factory.test.js
@@ -229,10 +229,6 @@ describe('The state url conversion factory', function () {
         });
 
         describe('The params to state translation', function () {
-            beforeEach(function () {
-
-            });
-
             it('translates empty params to an empty state', function () {
                 const state = stateUrlConverter.params2state({}, {});
                 expect(state).toEqual({});