--- a/modules/shared/services/environment/environment.factory.js
+++ b/modules/shared/services/environment/environment.factory.js
@@ -30,6 +30,8 @@
                 config.NAME = 'DEVELOPMENT';
         }
 
+        config.isDevelopment = () => config.NAME === 'DEVELOPMENT';
+
         return config;
     }
 })();