--- a/package.json
+++ b/package.json
@@ -4,8 +4,10 @@
   "devDependencies": {
     "autoprefixer": "^6.4.0",
     "babel-cli": "^6.18.0",
+    "babel-plugin-transform-object-rest-spread": "6.23.0",
     "babel-polyfill": "^6.16.0",
     "babel-preset-es2015": "^6.16.0",
+    "babel-preset-latest": "6.24.1",
     "bower": "1.8.0",
     "connect-livereload": "^0.5.4",
     "eslint": "^3.7.1",