Map zoom reducer speed increase from 20 ms to 1ms.

State frozen, also hitting performance by about 10ms.

URL resolution dp-link from 300ms to 3ms.

TEST URL:

http://localhost:8000/#?dte=gebieden%2Fstadsdeel%2F03630000000018%2F&mpb=topografie&mpz=9&mpv=52.3733157:4.9036913